# docs/swagger.yaml
basePath: /v1
definitions:
  github_com_serinew_core_internal_model_core.User:
    properties:
      createdAt:
        type: string
      email:
        type: string
      id:
        type: string
      name:
        type: string
      phone:
        type: string
      updatedAt:
        type: string
    type: object
  github_com_serinew_core_internal_model_core.UserAddress:
    properties:
      addr1:
        type: string
      addr2:
        type: string
      createdAt:
        type: string
      lat:
        type: number
      lng:
        type: number
      userId:
        type: string
      zipCode:
        type: integer
    type: object
  github_com_serinew_core_internal_model_core.UserProfile:
    properties:
      nickName:
        type: string
      profileImageUrl:
        type: string
      updatedAt:
        type: string
      userId:
        type: string
      userName:
        type: string
    type: object
  github_com_serinew_core_internal_types.ErrorDoc:
    properties:
      message:
        type: string
      status:
        type: string
      statusCode:
        type: integer
    type: object
  github_com_serinew_core_internal_types.SignChangePasswordReq:
    properties:
      newPassword:
        type: string
      oldPassword:
        type: string
      userId:
        type: string
    required:
    - newPassword
    - oldPassword
    - userId
    type: object
  github_com_serinew_core_internal_types.SignLoginAdmin:
    properties:
      authority:
        type: string
    type: object
  github_com_serinew_core_internal_types.SignLoginData:
    properties:
      address:
        items:
          $ref: '#/definitions/github_com_serinew_core_internal_model_core.UserAddress'
        type: array
      admin:
        $ref: '#/definitions/github_com_serinew_core_internal_types.SignLoginAdmin'
      createdAt:
        type: string
      email:
        type: string
      id:
        type: string
      name:
        type: string
      phone:
        type: string
      profile:
        $ref: '#/definitions/github_com_serinew_core_internal_model_core.UserProfile'
      updatedAt:
        type: string
    type: object
  github_com_serinew_core_internal_types.SignLoginReq:
    properties:
      password:
        type: string
      username:
        type: string
    required:
    - password
    - username
    type: object
  github_com_serinew_core_internal_types.SignUpReq:
    properties:
      addr1:
        maxLength: 256
        type: string
      addr2:
        type: string
      email:
        type: string
      lat:
        type: number
      lng:
        type: number
      name:
        type: string
      password:
        type: string
      phone:
        type: string
      username:
        type: string
      zipCode:
        type: integer
    required:
    - addr1
    - email
    - name
    - password
    - phone
    - username
    - zipCode
    type: object
  github_com_serinew_core_internal_types.SuccessEnvelopeMsg:
    properties:
      message:
        type: string
      status:
        type: string
      statusCode:
        type: integer
    type: object
  github_com_serinew_core_internal_types.SuccessEnvelopeSignLogin:
    properties:
      data:
        $ref: '#/definitions/github_com_serinew_core_internal_types.SignLoginData'
      message:
        type: string
      status:
        type: string
      statusCode:
        type: integer
    type: object
  github_com_serinew_core_internal_types.SuccessEnvelopeSignLoginList:
    properties:
      count:
        type: integer
      data:
        items:
          $ref: '#/definitions/github_com_serinew_core_internal_types.SignLoginData'
        type: array
      message:
        type: string
      status:
        type: string
      statusCode:
        type: integer
    type: object
  github_com_serinew_core_internal_types.SuccessEnvelopeStr:
    properties:
      data:
        type: string
      message:
        type: string
      status:
        type: string
      statusCode:
        type: integer
    type: object
  github_com_serinew_core_internal_types.SuccessEnvelopeUser:
    properties:
      data:
        $ref: '#/definitions/github_com_serinew_core_internal_model_core.User'
      message:
        type: string
      status:
        type: string
      statusCode:
        type: integer
    type: object
info:
  contact: {}
  description: Serinew core HTTP API. Swagger UI(`/v1/docs`)는 accessToken 쿠키로 로그인한
    사용자 중 core.admin 에 등록된 계정만 접근 가능합니다.
  title: Core API
  version: "1.0"
paths:
  /:
    get:
      description: '`/v1` 라우트 그룹 확인'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/github_com_serinew_core_internal_types.SuccessEnvelopeStr'
      summary: V1 상태
      tags:
      - meta
  /sign:
    get:
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/github_com_serinew_core_internal_types.SuccessEnvelopeSignLogin'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/github_com_serinew_core_internal_types.ErrorDoc'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/github_com_serinew_core_internal_types.ErrorDoc'
      summary: 현재 세션
      tags:
      - sign
  /sign/login:
    post:
      consumes:
      - application/json
      parameters:
      - description: 요청 바디
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/github_com_serinew_core_internal_types.SignLoginReq'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/github_com_serinew_core_internal_types.SuccessEnvelopeSignLogin'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/github_com_serinew_core_internal_types.ErrorDoc'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/github_com_serinew_core_internal_types.ErrorDoc'
      summary: 로그인
      tags:
      - sign
  /sign/password:
    post:
      consumes:
      - application/json
      parameters:
      - description: 요청 바디
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/github_com_serinew_core_internal_types.SignChangePasswordReq'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/github_com_serinew_core_internal_types.SuccessEnvelopeMsg'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/github_com_serinew_core_internal_types.ErrorDoc'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/github_com_serinew_core_internal_types.ErrorDoc'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/github_com_serinew_core_internal_types.ErrorDoc'
      summary: 비밀번호 변경
      tags:
      - sign
  /sign/register:
    post:
      consumes:
      - application/json
      parameters:
      - description: 요청 바디
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/github_com_serinew_core_internal_types.SignUpReq'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/github_com_serinew_core_internal_types.SuccessEnvelopeUser'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/github_com_serinew_core_internal_types.ErrorDoc'
        "409":
          description: Conflict
          schema:
            $ref: '#/definitions/github_com_serinew_core_internal_types.ErrorDoc'
      summary: 회원 가입
      tags:
      - sign
  /users:
    get:
      parameters:
      - description: 페이지 (기본 1)
        in: query
        name: page
        type: integer
      - description: 페이지 크기 (기본 20, 최대 100)
        in: query
        name: pageSize
        type: integer
      - description: 검색어
        in: query
        name: search
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/github_com_serinew_core_internal_types.SuccessEnvelopeSignLoginList'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/github_com_serinew_core_internal_types.ErrorDoc'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/github_com_serinew_core_internal_types.ErrorDoc'
      security:
      - CookieAuth: []
      summary: 사용자 목록
      tags:
      - users
  /users/{userId}:
    get:
      parameters:
      - description: 사용자 UUID
        in: path
        name: userId
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/github_com_serinew_core_internal_types.SuccessEnvelopeSignLogin'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/github_com_serinew_core_internal_types.ErrorDoc'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/github_com_serinew_core_internal_types.ErrorDoc'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/github_com_serinew_core_internal_types.ErrorDoc'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/github_com_serinew_core_internal_types.ErrorDoc'
      security:
      - CookieAuth: []
      summary: 사용자 단건
      tags:
      - users
schemes:
- http
- https
securityDefinitions:
  CookieAuth:
    in: cookie
    name: accessToken
    type: apiKey
swagger: "2.0"